new shop and illustrations ref

+++ app/controllers/images_controller.rb
@@ -1,10 +1,12 @@
 class ImagesController < ApplicationController
-  before_filter :require_artist
   def destroy
     @product = Product.find(params[:product_id])
-    im = @product.images.find(params[:id])
-    im.destroy
-    @product.save
-    redirect_to edit_shop_product_path(@product.shop, @product)
+    @shop = Shop.find_by_name(params[:shop_id])
+    if check_artist?
+      im = @product.images.find(params[:id])
+      im.destroy
+      @product.save
+      redirect_to edit_shop_product_path(@product.shop, @product)
+    end
   end
 end
